+++ src/lib/components/organisms/NewsComponent.svelte
