+++ src/lib/index.js
@@ -1,6 +1,6 @@
 // Assets
 
 // General components
-export { default as Hero } from './components/molecules/Hero/Hero.svelte'
-export { default as Breadcrumb } from './components/molecules/Breadcrumb/Breadcrumb.svelte'
-export { default as Footer } from './components/organisms/Footer/Footer.svelte'
+export { default as Hero } from './components/organisms/Hero.svelte'
+export { default as Breadcrumb } from './components/molecules/Breadcrumb.svelte'
+export { default as Footer } from './components/organisms/Footer.svelte'
